requirements.txt
alembic==1.12.0
fastapi==0.103.1
fastapi-sqlalchemy==0.2.1

uvicorn==0.23.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
passlib==1.7.4

python-jose==3.3.0
python-multipart==0.0.6
